requirements.txt
streamlit>=1.40.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
openpyxl>=3.1.0
requests>=2.28.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
mplfinance>=0.12.0
yfinance>=0.2.40
pymaya>=0.1.0
edgartools>=4.0.0
riskfolio-lib>=7.0.0

# Gemini calls — used by scripts/run_recommendations.py and the in-app Run AI button.
# Required on Streamlit Cloud so the in-app "Run analysis →" works.
langchain-google-genai>=2.0.0
